+++ AlchemistCode/Cards/Common/CausticStrike.cs
@@ -15,7 +15,7 @@ public class CausticStrike : AlchemistCard
 
     public CausticStrike() : base(1, CardType.Attack, CardRarity.Common, TargetType.AnyEnemy)
     {
-        WithDamage(6, 2);
+        WithDamage(5, 2);
         WithVar(new FermentVar("Poison", 2, perTurn: 2).WithUpgrade(1));
         WithKeyword(CardKeyword.Retain);
         WithTip(typeof(PoisonPower));