+++ CHANGELOG.md
@@ -7,6 +7,22 @@ mod).
 
 ## [Unreleased]
 
+### Changed
+
+- Changed Ferment cards to show the fermented turn count only during combat
+- Changed Tempered: Block 8 (12) -> 10 (13), and additional Block for each card in your Exhaust Pile 3 -> 1 (2)
+- Nerfed Bloom: fermentation 3 (4) -> 2 (3)
+- Nerfed Caustic Strike damage: 6 (8) -> 5 (7)
+- Nerfed Eureka: it now Exhausts
+- Nerfed Puff Up: additional Block for each turn fermented 4 (6) -> 3 (5)
+- Nerfed Ripening: cost increased from 1 -> 2 (1), and Ripening+ no longer adds a random Mix into your Hand each turn
+- Nerfed Untended: drawn Ferment cards ferment 1 instead of 2
+- Reworked Brace: "Gain 7 (9) Block. Whenever you are attacked this turn, apply 2 (3) Poison to the attacker."
+
+### Fixed
+
+- Compound Mix cards now properly show damage and Block modified by Strength, Dexterity and debuffs
+
 ## [0.14.11] - 2026-09-10
 
 ### Added